# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Catalog API",
        "version": "1.0.0",
        "description": "Lightweight Music Catalog Microservice optimized for STB devices. This API provides access to a Spotify-based music catalog with track information, genre filtering, and AI-powered recommendations.",
        "contact": {
            "name": "Anggota B - Sistem Terintegrasi"
        }
    },
    "servers": [
        {
            "url": "https://ibay.ibayderikfariqalghanzaka.my.id",
            "description": "Production server"
        },
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        }
    ],
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "API Key untuk autentikasi. Gunakan key dari file .env Anda"
            }
        },
        "schemas": {
            "TrackListItem": {
                "type": "object",
                "description": "Format ringkas untuk list tracks (pagination)",
                "properties": {
                    "track_id": {
                        "type": "string",
                        "description": "Spotify Track ID",
                        "example": "3322ArxAq7wCSZI4fF77Q0"
                    },
                    "track_name": {
                        "type": "string",
                        "description": "Nama lagu",
                        "example": "Please Don't Say You Love Me"
                    },
                    "artists": {
                        "type": "string",
                        "description": "Nama artis",
                        "example": "Gabrielle Aplin"
                    },
                    "track_genre": {
                        "type": "string",
                        "description": "Genre musik",
                        "example": "acoustic"
                    }
                }
            },
            "TrackDetail": {
                "type": "object",
                "description": "Format lengkap track dengan audio features (untuk Gemini AI)",
                "properties": {
                    "track_id": {
                        "type": "string",
                        "description": "Spotify Track ID",
                        "example": "3322ArxAq7wCSZI4fF77Q0"
                    },
                    "track_name": {
                        "type": "string",
                        "description": "Nama lagu",
                        "example": "Please Don't Say You Love Me"
                    },
                    "artists": {
                        "type": "string",
                        "description": "Nama artis",
                        "example": "Gabrielle Aplin"
                    },
                    "album_name": {
                        "type": "string",
                        "description": "Nama album",
                        "example": "English Rain",
                        "nullable": true
                    },
                    "track_genre": {
                        "type": "string",
                        "description": "Genre musik",
                        "example": "acoustic"
                    },
                    "audio_features": {
                        "type": "object",
                        "description": "Fitur audio untuk analisis AI",
                        "properties": {
                            "energy": {
                                "type": "number",
                                "format": "float",
                                "description": "Tingkat energi (0.0 - 1.0)",
                                "minimum": 0,
                                "maximum": 1,
                                "example": 0.45
                            },
                            "valence": {
                                "type": "number",
                                "format": "float",
                                "description": "Tingkat kebahagiaan/positivity (0.0 - 1.0)",
                                "minimum": 0,
                                "maximum": 1,
                                "example": 0.62
                            },
                            "acousticness": {
                                "type": "number",
                                "format": "float",
                                "description": "Tingkat akustik (0.0 - 1.0)",
                                "minimum": 0,
                                "maximum": 1,
                                "example": 0.85
                            },
                            "tempo": {
                                "type": "number",
                                "format": "float",
                                "description": "BPM (Beats Per Minute)",
                                "example": 120.5
                            }
                        }
                    }
                }
            },
...
            "Pagination": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "description": "Halaman saat ini",
                        "example": 1
                    },
                    "limit": {
                        "type": "integer",
                        "description": "Jumlah item per halaman",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "description": "Total jumlah records",
                        "example": 3121
                    },
                    "totalPages": {
                        "type": "integer",
                        "description": "Total jumlah halaman",
                        "example": 313
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "description": "Pesan error",
                        "example": "Track not found"
                    }
                }
            }
        }
    },
...
    "paths": {
        "/music": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "List semua tracks dengan pagination",
                "description": "Mengembalikan daftar track musik dengan pagination. Mendukung filtering berdasarkan genre.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah item per halaman (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "minimum": 1,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar tracks berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TrackListItem"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - API Key invalid atau tidak ada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "success": false,
                                    "error": "Invalid API Key"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/music/{id}": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Get track berdasarkan ID",
                "description": "Mengembalikan detail lengkap track termasuk audio features. ID harus berupa Spotify Track ID.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Spotify Track ID",
                        "schema": {
                            "type": "string"
                        },
                        "example": "3322ArxAq7wCSZI4fF77Q0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail track berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TrackDetail"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Track tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "success": false,
                                    "error": "Track not found"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/music/batch": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Get multiple tracks sekaligus",
                "description": "Mengembalikan beberapa track dalam satu request menggunakan comma-separated track IDs. Maksimal 50 tracks per request.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "required": true,
                        "description": "Comma-separated Spotify Track IDs",
                        "schema": {
                            "type": "string"
                        },
                        "example": "3322ArxAq7wCSZI4fF77Q0,4LGF2tDg3878bs0mQPByZ4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch tracks berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TrackDetail"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - parameter IDs tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "no_ids": {
                                        "summary": "Tidak ada IDs",
                                        "value": {
                                            "success": false,
                                            "error": "No track IDs provided"
                                        }
                                    },
                                    "too_many": {
                                        "summary": "Terlalu banyak IDs",
                                        "value": {
                                            "success": false,
                                            "error": "Maximum 50 tracks per batch request"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/music/search": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Cari tracks berdasarkan genre",
                "description": "Mencari tracks berdasarkan genre dengan pagination",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "genre",
                        "in": "query",
                        "required": true,
                        "description": "Nama genre untuk dicari",
                        "schema": {
                            "type": "string"
                        },
                        "example": "acoustic"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah item per halaman (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil pencarian",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TrackListItem"
                                            }
                                        },
                                        "filter": {
                                            "type": "object",
                                            "properties": {
                                                "genre": {
                                                    "type": "string",
                                                    "example": "acoustic"
                                                }
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 15
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - parameter genre wajib",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "success": false,
                                    "error": "Genre parameter is required"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/music/recommendations": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Get rekomendasi musik",
                "description": "Mendapatkan rekomendasi track berdasarkan genre dan mood (menggunakan audio features)",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Filter berdasarkan genre",
                        "schema": {
                            "type": "string"
                        },
                        "example": "pop"
                    },
                    {
                        "name": "mood",
                        "in": "query",
                        "description": "Filter berdasarkan mood",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "energetic",
                                "calm",
                                "happy",
                                "sad"
                            ]
                        },
                        "example": "energetic"
                    },
                    {
                        "name": "minEnergy",
                        "in": "query",
                        "description": "Energy minimum (0.0 - 1.0)",
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0,
                            "maximum": 1
                        }
                    },
                    {
                        "name": "maxEnergy",
                        "in": "query",
                        "description": "Energy maksimum (0.0 - 1.0)",
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0,
                            "maximum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah rekomendasi (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rekomendasi berhasil didapat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TrackDetail"
                                            }
                                        },
                                        "filters": {
                                            "type": "object",
                                            "description": "Filter yang digunakan"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 20
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}